requirements.txt
# Robot Framework và extensions
robotframework==6.1.1
robotframework-seleniumlibrary==6.2.0
robotframework-appiumlibrary==2.1.0
robotframework-requests==0.9.5
robotframework-pabot==2.16.0

# Appium và mobile testing
Appium-Python-Client==3.1.1
selenium==4.15.2

# Báo cáo và logging
allure-robotframework==2.13.2
allure-pytest==2.13.2
allure-python-commons==2.13.2

# Data handling
PyYAML==6.0.1
pandas==2.1.3
openpyxl==3.1.2

# Testing utilities
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1

# Image processing cho screenshot
Pillow==10.1.0

# Configuration management
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# JSON handling
jsonschema==4.20.0

# Logging
colorlog==6.8.0

# Parallel execution
multiprocessing-logging==0.3.4 